move all red team components to app/redteaming folder
