+++ app/redteaming/(fullscreen)/components/attackModuleSelector.tsx
