+++ app/redteaming/(fullscreen)/components/chatbox-slide-box-sizes.ts
