+++ app/redteaming/(fullscreen)/components/chatbox-slide-layout.tsx
