+++ app/redteaming/(fullscreen)/components/chatbox.tsx
