+++ app/redteaming/(fullscreen)/components/color-coded-template.tsx
