+++ app/redteaming/(fullscreen)/components/contextStrategiesList.tsx
