+++ app/redteaming/(fullscreen)/components/prompt-box.tsx
@@ -5,7 +5,7 @@ import { TextArea } from '@/app/components/textArea';
 import { Tooltip, TooltipPosition } from '@/app/components/tooltip';
 import useOutsideClick from '@/app/hooks/use-outside-click';
 import { debounce } from '@/app/lib/throttle';
-import useChatboxesPositionsUtils from '@/app/views/redteaming/hooks/useChatboxesPositionsUtils';
+import useChatboxesPositionsUtils from '@/app/redteaming/(fullscreen)/hooks/useChatboxesPositionsUtils';
 import { colors } from '@/app/views/shared-components/customColors';
 import { toggleDarkMode, useAppDispatch } from '@/lib/redux';
 import {
