+++ app/redteaming/(fullscreen)/components/prompt-bubble-info.tsx
