+++ app/redteaming/(fullscreen)/components/promptAndResponseBubble.tsx
