+++ app/redteaming/(fullscreen)/components/promptTemplatesList.tsx
