+++ app/redteaming/(fullscreen)/components/selectedOptionPill.tsx
