+++ app/redteaming/(fullscreen)/components/slash-commands.ts
