+++ app/redteaming/(fullscreen)/hooks/useChatboxesPositionsUtils.ts
