+++ app/redteaming/(others)/sessions/new/enums.ts
