+++ app/redteaming/(others)/sessions/new/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RedteamNewSessionFlow } from '@/app/views/redteaming/redteamNewSessionFlow';
+import { RedteamNewSessionFlow } from './redteamNewSessionFlow';
 
 export default function BenchmarkNewSessionFlowPage() {
   return <RedteamNewSessionFlow />;
