+++ app/redteaming/(others)/sessions/new/redteamNewSessionFlow.tsx
@@ -5,6 +5,7 @@ import { Icon, IconName } from '@/app/components/IconSVG';
 import { Button, ButtonType } from '@/app/components/button';
 import SimpleStepsIndicator from '@/app/components/simpleStepsIndicator';
 import { NewEndpointForm } from '@/app/endpoints/(edit)/newEndpointForm';
+import { AttackModuleSelectView } from '@/app/redteaming/(fullscreen)/components/attackModuleSelector';
 import { EndpointSelectVew } from '@/app/views/models-management/endpointsSelector';
 import { colors } from '@/app/views/shared-components/customColors';
 import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
@@ -18,7 +19,6 @@ import {
   useAppDispatch,
   useAppSelector,
 } from '@/lib/redux';
-import { AttackModuleSelectView } from './components/attackModuleSelector';
 import { RedteamingNewSessionViews } from './enums';
 import { RedteamRunForm } from './redteamRunForm';
